% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.probes.R
\name{collapse.probes}
\alias{collapse.probes}
\title{collapse eSet object by non-unique row feature IDs}
\usage{
collapse.probes(
  eset,
  fdat.id = "SYMBOL",
  method = c("single.mad", "summarize.mean", "summarize.median", "summarize.max",
    "summarize.min")
)
}
\arguments{
\item{fdat.id}{feature id to collapse on}

\item{method}{= "single" or "summarize" followed by "." followed by "mean", "median", "max" or "min", e.g. "single.mean"
"single" chooses one probe for ambiguous probes with maximum of summary function across row, 
"summarize" aggregates summary metric across multiple rows for each column index}

\item{eSet}{expression set object}
}
\description{
collapse eSet object by non-unique row feature IDs
}
