% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{eSet.brca.100}
\alias{eSet.brca.100}
\title{Example dataset
TCGA breast cancer RNA-seq expression set (mad filtered 100 genes), log2 transformed}
\format{expression set object}
\usage{
data(eSet.brca.100)
}
\description{
Example dataset
TCGA breast cancer RNA-seq expression set (mad filtered 100 genes), log2 transformed
}

