% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggheatmap.R
\docType{data}
\name{eSet1}
\alias{eSet1}
\title{Example dataset #1
TCGA breast cancer RNA-seq expression set (mad filtered 500 genes), log2 transformed}
\format{expression set object}
\usage{
data(eSet1)
}
\description{
Example dataset #1
TCGA breast cancer RNA-seq expression set (mad filtered 500 genes), log2 transformed
}

