% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/broad.file.formats.R
\docType{class}
\name{gctdata-class}
\alias{gctdata-class}
\title{gctdata}
\description{
Class \code{gctdata} represents a gene expression microarray data in the Broad '.gct' format
}
\section{Slots}{

\describe{
\item{\code{signal}}{is the [m x n] gene-by-sample matrix of expression levels}

\item{\code{description}}{is an m-sized character vector of genes' (or probesets') descriptions}
}}

