% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gspData}
\alias{gspData}
\title{Example dataset
gene set projection input data example}
\format{expression set object}
\usage{
data(gspData)
}
\value{
\itemize{
 \item{"gsp.eSet"}{an eSet object}
 \item{"gsp.GeneSet"}{an S4 geneset object defined by \code{GeneSet-class}}
}
}
\description{
Example dataset
gene set projection input data example
}

