% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hyper.enrichment.R
\name{hyper.enrichment}
\alias{hyper.enrichment}
\title{hyper.enrichment}
\usage{
hyper.enrichment(drawn, categories,
  ntotal = length(unique(unlist(categories))), min.drawsize = 4,
  mht = TRUE, verbose = TRUE)
}
\arguments{
\item{drawn}{one or more sets of 'drawn' items (e.g., genes). Basically, a list of signatures.}

\item{categories}{list of gene sets (e.g., MSigDB c2)}

\item{ntotal}{background population, i.e., the total no. of items from which items are supposed to have been drawn}

\item{min.drawsize}{min no. of drawn items that must be among categories' items}

\item{mht}{correct for multiple hypothesis testing across multiple 'draws'}
}
\value{
a data.frame with rows indexed by the signature(s) tested
}
\description{
\code{hyper.enrichment} carry out set enrichment test based on hyper-geometric distribution
}
\examples{
# load objects hyperSig (a list of signatures) and hyperGsets (a GeneSet object)
# and run hyper-enrichment test

data(hyper)
print(hyper)
hyperE<-hyper.enrichment(drawn=hyperSig,categories=geneSet(hyperGsets),ntotal=10000)
}

