% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hyper}
\alias{hyper}
\title{Example dataset
hyperenrichment input data example}
\format{expression set object}
\usage{
data(hyper)
}
\value{
\itemize{
 \item{"hyperGsets"}{an S4 geneset object defined by \code{GeneSet-class}}
 \item{"hyperSig"}{a list of character vectors (geneSets)}
}
}
\description{
Example dataset
hyperenrichment input data example
}

