% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GeneSet.R
\docType{methods}
\name{length,GeneSet-method}
\alias{length,GeneSet-method}
\title{length}
\usage{
\S4method{length}{GeneSet}(x)
}
\description{
method \code{length} returns the number of genesets contained in the object
}

