% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{\code{my_add} add two numbers and print it}
\usage{
my_add(x, y)
}
\arguments{
\item{x}{first number}

\item{y}{second number}
}
\description{
\code{my_add} add two numbers and print it
}

