% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nn.analysis.R
\name{nn.analysis}
\alias{nn.analysis}
\title{nn.analysis}
\usage{
nn.analysis(dat, y = NULL, y.i = NULL, score = c("pearson", "spearman",
  "euclid"), nperm = 0, control = NULL, seed = NULL, verbose = F,
  smooth = 1)
}
\arguments{
\item{dat}{NxM genes-by-sample matrix ('ExpressionSet' or 'resdata' or 'gctdata' object)}

\item{y}{M-tuple vector to use as the prototype profile}

\item{y.i}{alternatively, specify the row index in dat where to find the prototype profile}

\item{score}{similarity measure, must be one of 'pearson', 'spearman' or 'euclid'}

\item{nperm}{number of permutation iterations (default 0)}

\item{control}{constrained permutations (control for confounding factor)}

\item{seed}{}

\item{verbose}{}

\item{smooth}{smoother to prevent 0 p-values (default 1)}
}
\value{
list object with attributes 'pval' and 'nperm' corresponding to nn analysis permutation results
}
\description{
\code{nn.analysis} rank rows (genes) by their 'similarity' to a prototype profile
}
\examples{
rm(list=ls())
 CBMDEV <- Sys.getenv('CBMDEV')
 source(paste(CBMDEV,'R/source.nn.analysis.R',sep='/'))
 require(Biobase)
 DAT <- readRDS(paste(CBMDEV,'data/TCGA_OSCC_mRNA.rds',sep='/'))
 DAT1 <- DAT[order(apply(exprs(DAT),1,mad),decreasing=TRUE)[1:1000],]
 ## passing the actual vector as argument
 ##
 NNout1 <- nn.analysis(dat=DAT1,y=exprs(DAT1)[1,],score='pearson',nperm=1000,seed=123,verbose=TRUE,smooth=1)
 head(NNout1$pval)
 ## passing the index (within the matrix) as argument
 ##
 NNout2 <- nn.analysis(dat=DAT1,y.i=1,score='pearson',nperm=1000,seed=123,verbose=TRUE,smooth=1)
 head(NNout2$pval)
 all.equal(NNout1$pval,NNout2$pval) # TRUE
 ## controlling for a confounder (only discrete)
 ##
 CTL <- pData(DAT1)[,'treatment']
 NNout3 <- nn.analysis(dat=DAT1,y.i=1,score='pearson',nperm=1000,seed=123,control=CTL,verbose=TRUE,smooth=1)
 head(NNout3$pval)
}

