% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oligo_wrapper.R
\name{oligo_wrapper}
\alias{oligo_wrapper}
\title{wrapper script for microarray gene ST chips processing into expression set object}
\usage{
oligo_wrapper(indir, outdir, header)
}
\arguments{
\item{indir}{input directory}

\item{outdir}{output directory}

\item{header}{of outfile}
}
\value{
the normalized expressionSet object
}
\description{
wrapper script for microarray gene ST chips processing into expression set object
}
\examples{
#running example
#indir<-"/Users/amyli/Desktop/temp_dir/example_script/raw"
#outdir<-"/Users/amyli/Desktop/temp_dir/example_script/processed"
#dat<-oligo_wrapper(indir = indir, outdir = outdir, header = "my_eset")


}
