% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.xlsx.R
\name{read.xlsx}
\alias{read.xlsx}
\title{\code{read.xlsx} read from .xlsx}
\usage{
read.xlsx(file, java.param = "-Xmx3g")
}
\arguments{
\item{file}{the xlsx file to read from, may contain multiple sheets}

\item{java.param}{max java memory, default =  "-Xmx3g"}
}
\description{
\code{read.xlsx} read from .xlsx
}

