% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/broad.file.formats.R
\docType{class}
\name{resdata-class}
\alias{resdata-class}
\title{resdata}
\description{
Class \code{resdata} represents a gene expression microarray data in the Broad '.res' format
}
\section{Slots}{

\describe{
\item{\code{signal}}{is the [m x n] gene-by-sample matrix of expression levels}

\item{\code{calls}}{is the [m x n] gene-by-sample matrix of P/M/A calls}

\item{\code{description}}{is an m-sized character vector of genes' (or probesets') descriptions}

\item{\code{scale}}{is the scaling factor used with each sample (if any)}
}}

