% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_wrapper.R
\name{run_limma}
\alias{run_limma}
\title{wrapper for running limma with additional columns attached}
\usage{
run_limma(
  eset,
  treatment,
  cond,
  control,
  sort.by = "adj.P.Val",
  decreasing = FALSE,
  cutoff = NA,
  verbose = TRUE
)
}
\arguments{
\item{eset}{expression set object with gene expression}

\item{treatment}{column name in pData(eset) specifying the conditions for differential expression}

\item{cond}{value in treatment specifying the positive condition}

\item{control}{value in treatment specifying the control condition}

\item{sort.by}{column to sort by}

\item{decreasing}{direction for sorting}

\item{cutoff}{threshold value for keeping results}

\item{verbose}{whether to display additional columns in report table}
}
\description{
wrapper for running limma with additional columns attached
}
\examples{
data(eSet.brca.100)
eset<-eSet.brca.100
eset<-eset[, eset$ER_status \%in\% c("Positive", "Negative")]
res<-run_limma(eset = eset, treatment = "ER_status", cond = "Positive",
	control = "Negative", verbose = TRUE, sort.by = "adj.P.Val", decreasing = FALSE, cutoff = 0.05)
}
