% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/save.xlsx.R
\name{save.xlsx}
\alias{save.xlsx}
\title{\code{save.xlsx} save to .xlsx}
\usage{
save.xlsx(x, f.dir = NA, f.header = "", f.name, java.param = "-Xmx3g")
}
\arguments{
\item{x}{R object to save, must be a named list of data.frames, row limit =65535 for each data frame}

\item{f.dir}{directory to save to, default = NA, current working directory}

\item{f.header}{file header}

\item{f.name}{file name (appended to file header)}
}
\description{
\code{save.xlsx} save to .xlsx
}

