% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga.subset.400g.200s}
\alias{tcga.subset.400g.200s}
\title{Example dataset
TCGA breast cancer RNA-seq expression set (mad filtered 400 genes), log2 transformed, 200 samples, has subtype info}
\format{expression set object}
\usage{
data(tcga.subset.400g.200s.rda)
}
\description{
Example dataset
TCGA breast cancer RNA-seq expression set (mad filtered 400 genes), log2 transformed, 200 samples, has subtype info
}
