% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.eset.R
\name{to.eSet}
\alias{to.eSet}
\title{construct an eSet object}
\usage{
to.eSet(mat, pdat, fdat)
}
\arguments{
\item{mat}{expression matrix}

\item{pdat}{phenoData data frame}

\item{fdat}{featureData data frame}
}
\description{
construct an eSet object
}

