% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/variationFilter.R
\name{variationFilter}
\alias{variationFilter}
\title{variationFilter}
\usage{
variationFilter(dat, score = c("mad", "sd", "cv"), dir = c("top", "bottom"),
  transform = c("none", "log2", "exp2", "log", "exp"), ngenes = NULL,
  min.score = NULL, min.log = 1, rnd = 4, do.plot = FALSE, pch = ".",
  lgnd.coord = 1, do.log = NULL, qnt.lev = 0.5, min.qnt = -Inf,
  no.affx = FALSE, verbose = TRUE)
}
\arguments{
\item{dat}{an expressionSet object}

\item{score}{type of variation measure to use (can be 'mad', 'sd', or 'cv')}

\item{dir}{direction of the gene ranking ('top': highest varying genes; 'bottom': lowest varying genes)}

\item{transform}{transform the data before measuring variation}

\item{ngenes}{number of top (or bottom) genes to extract}

\item{do.plot}{plot the center vs. scale plot with selected genes highlighted}
}
\value{
the filtered expressionSet object
}
\description{
\code{variationFilter} subset an expression dataset by extracting genes based on their level of variation
}
\examples{
# select the top 100 genes by median absolute deviation (mad) and also generate plot

# Use example data, for data set information: ?eSet.brca.100
data(eSet.brca.100)

variationFilter(eSet.brca.100,score='mad',ngenes=50,do.plot=TRUE)
}

