% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geneSetProjection.R
\name{geneSetProjection}
\alias{geneSetProjection}
\title{geneSetProjection}
\usage{
geneSetProjection(dat, pairing, GS, method = c("ks", "logistic", "median",
  "mean"), collapse = T, weighted = T, absolute = F, min.gset = 5,
  verbose = T)
}
\arguments{
\item{dat}{ExpressionSet data object ( require(biobase) )}

\item{pairing}{list with pairing between treatment and control (see format below)}

\item{method}{one of {"ks","logistic","median","mean"} (only ks implemented so far)}

\item{collapse}{collapse multiple replicates into a single output vector (TRUE)}

\item{weighted}{gsea-like weighting of KS score (TRUE)}

\item{absolute}{use absolute values when calculating the KS score ignoring up and down (FALSE)}

\item{verbose}{verbosity on/off (TRUE)}

\item{gsets}{GeneSet data object (see GeneSet.R)}
}
\value{
an expression set
}
\description{
\code{geneSetProjection} map a gene-level dataset to a
geneset-level dataset based on control-treatment pairing (i.e., the
ks score is done with respect to the 'control vs. treatment'
phenotype).  This function is ideal for 'time-series' data of
response to treatment, where for each time point there are several
replicates for both the control and the treated.
}

