% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_wrapper.R
\name{run_deseq}
\alias{run_deseq}
\title{Run DESeq2 on eset object}
\usage{
run_deseq(eset, class_id, control, treatment, verbose = TRUE,
  sortByPValue = TRUE, FDRcutoff = NULL)
}
\arguments{
\item{eset}{expression set object with gene expression}

\item{class_id}{column name in pData(eset) specifying the conditions for differential expression}

\item{control}{value in treatment specifying the control condition}

\item{treatment}{value in treatment specifying the positive condition}

\item{verbose}{add additional information and perform filtering}

\item{sortByPValue}{whether to sort results by nominal p-value, if verbose = TRUE}

\item{FDRcutoff}{FDR threshold for filtering results, if verbose = TRUE}

\item{control}{value in treatment specifying the control condition}
}
\description{
Run DESeq2 on eset object
}

